fix(state): stage compaction rows before atomic cutover
The lease-backed in-place compaction path serialized replacement rows, fed both FTS indexes, archived the old transcript, cloned the concurrent tail, and updated counters inside one BEGIN IMMEDIATE. On large shared state.db files that single writer hold reached 25-140 seconds, long enough for unrelated gateway turns to exhaust transcript-write patience and stop as session storage busy.

Build replacement rows under a hidden internal stage session in row- and byte-bounded transactions, then use one short lease-fenced cutover to archive the old active rows, move the complete staged set onto the durable session, preserve tail_count rewind semantics, clone the post-watermark concurrent tail, merge model config, and update counters. A sibling process can write between stage chunks, while readers see either the complete old view or the complete compacted view.

Failure cleanup is bounded and exact-stage scoped. A later proven lock owner reclaims crash residue before staging. Calls without a compression holder keep the historical single-transaction implementation so the micro/prune ownership work in #77861 and #80245 remains separate.

Move the existing schema mixin byte-for-byte behind a compatibility loader so this implementation lands without adding to the grandfathered hermes_state.py godfile; the public hermes_state_schema module remains the single mutable namespace.

Fixes #90173
Related #97963
Reported-by: LeonardoLGDS
